Cleaned diff
