+++ src/test/resources/outputs/NestedNullTest.java
@@ -1,0 +1,17 @@
+public class NestedNullTest {
+    public boolean checkNull(String str) {
+        return str != null;
+    }
+
+    public void test() {
+        String str = "Hello World";
+
+        if (str != null) {
+            ;
+        }
+
+        if (!(str != null)) {
+            ;
+        }
+    }
+}
